Lock dates in time clock edit dialog
